`Refactor import paths for RatingStars component to use new location in components/ui directory`
